pyproject.toml
[tool.poetry]
name = "langchain_udemy"
version = "0.1.0"
description = ""
authors = ["masato-210831 <b86500b@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
gradio = "3.31.0"
python-dotenv = "1.0.0"
langchain = "0.0.172"
openai = "0.27.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
